// data/github/current/ui/packages/copilot-pr-review-banner/__generated__/useLoadTreeComparisonQuery.graphql.ts
export type useLoadTreeComparisonQuery$data = {
  readonly node: {
    readonly additions?: number;
    readonly author?: {
      readonly login: string;
    } | null | undefined;
    readonly baseRefOid?: any;
    readonly baseRepository?: {
      readonly databaseId: number | null | undefined;
    } | null | undefined;
    readonly closed?: boolean;
    readonly deletions?: number;
    readonly headRefOid?: any;
    readonly headRepository?: {
      readonly databaseId: number | null | undefined;
    } | null | undefined;
    readonly isDraft?: boolean;
  } | null | undefined;
  readonly viewer: {
    readonly isCopilotDotcomChatEnabled: boolean;
    readonly login: string;
  };
};